+++ src/hooks/postPageHooks/usePostPage.ts
@@ -2,11 +2,21 @@ import { useRouter } from "next/router";
 import { useQuery } from "react-query";
 import { getApi } from "../../api/postpage";
 
-export function useEachPost(boardType: string) {
+const title: { [key: string]: string } = {
+  전체: "entireboard",
+  자유: "F",
+  유머: "H",
+  보스: "B",
+  직업: "I",
+};
+
+export function useEachPost() {
   const router = useRouter();
   const pageNum = parseInt(router.query.page as string) || 1;
+  const boardTitle = router.query.post as string;
+  const boardKey = title[boardTitle];
 
-  return useQuery([boardType, pageNum], () => getApi(boardType, pageNum), {
+  return useQuery([boardKey, pageNum], () => getApi(boardKey, pageNum), {
     refetchOnWindowFocus: false,
   });
 }
